* using generic [u]intptr_t instead of the creapy definitions in Cudd package * removed some warnings concering NULL
* TODO: send a patch to the Cudd guys, don't ever want to patch the package again!

+++ genet/libs/cudd/cudd/cuddBddCorr.c
@@ -476,9 +476,9 @@ CorrelHash(
 
     entry = (HashEntry *) key;
 #if SIZEOF_VOID_P == 8 && SIZEOF_INT == 4
-    val = ((int) ((long long)entry->f))*997 + ((int) ((long long)entry->g));
+    val = ((int) ((intptr_t)entry->f))*997 + ((int) ((intptr_t)entry->g));
 #else
-    val = ((int) entry->f)*997 + ((int) entry->g);
+    val = ((int) ((intptr_t)entry->f))*997 + ((int) ((intptr_t)entry->g));
 #endif
 
     return ((val < 0) ? -val : val) % modulus;
