+++ genet/libs/cudd/st/st.c
@@ -19,6 +19,7 @@
 
 #include "util.h"
 #include "st.h"
+#include <stdint.h>
 
 /*---------------------------------------------------------------------------*/
 /* Constant declarations                                                     */
@@ -46,7 +47,7 @@ static char rcsid[] UTIL_UNUSED = " $Id: st.c,v 1.11 2004/02/11 22:31:59 fabio E
 
 #define ST_NUMCMP(x,y) ((x) != (y))
 
-#define ST_NUMHASH(x,size) (ABS((long long)x)%(size))
+#define ST_NUMHASH(x,size) (ABS((intptr_t)x)%(size))
 
 #if SIZEOF_VOID_P == 8
 #define st_shift 3
@@ -54,7 +55,7 @@ static char rcsid[] UTIL_UNUSED = " $Id: st.c,v 1.11 2004/02/11 22:31:59 fabio E
 #define st_shift 2
 #endif
 
-#define ST_PTRHASH(x,size) ((unsigned int)((unsigned long long)(x)>>st_shift)%size)
+#define ST_PTRHASH(x,size) ((unsigned int)((uintptr_t)(x)>>st_shift)%size)
 
 #define EQUAL(func, x, y) \
     ((((func) == st_numcmp) || ((func) == st_ptrcmp)) ?\
@@ -319,7 +320,7 @@ st_lookup(st_table *table, void *key, void *value)
 
 ******************************************************************************/
 int
-st_lookup_int(st_table *table, void *key, int *value)
+st_lookup_int(st_table *table, void *key, intptr_t *value)
 {
     int hash_val;
     st_table_entry *ptr, **last;
@@ -331,8 +332,8 @@ st_lookup_int(st_table *table, void *key, int *value)
     if (ptr == NIL(st_table_entry)) {
 	return 0;
     } else {
-	if (value != NIL(int)) {
-	    *value = (int) (long long) ptr->record;
+	if (value != NIL(intptr_t)) {
+	    *value = (intptr_t) ptr->record;
 	}
 	return 1;
     }
@@ -687,7 +688,7 @@ st_delete_int(st_table *table, void *keyp, int *value)
     }
 
     *last = ptr->next;
-    if (value != NIL(int)) *value = (int) (long long) ptr->record;
+    if (value != NIL(int)) *value = (int) (intptr_t) ptr->record;
     *(char **)keyp = ptr->key;
     FREE(ptr);
     table->num_entries--;
@@ -969,7 +970,7 @@ st_gen_int(st_generator *gen, void *key_p, int *value_p)
     }
     *(char **)key_p = gen->entry->key;
     if (value_p != NIL(int)) {
-   	*value_p = (int) (long long) gen->entry->record;
+   	*value_p = (int) (intptr_t) gen->entry->record;
     }
     gen->entry = gen->entry->next;
     return 1;
