+++ genet/libs/cudd/st/st.h
@@ -204,7 +204,7 @@ extern st_table *st_init_table_with_params (ST_PFICPCP, ST_PFICPI, int, int, dou
 extern st_table *st_init_table (ST_PFICPCP, ST_PFICPI); 
 extern void st_free_table (st_table *);
 extern int st_lookup (st_table *, void *, void *);
-extern int st_lookup_int (st_table *, void *, int *);
+extern int st_lookup_int (st_table *, void *, intptr_t *);
 extern int st_insert (st_table *, void *, void *);
 extern int st_add_direct (st_table *, void *, void *);
 extern int st_find_or_add (st_table *, void *, void *);
